requirements.txt
pytest
pylint
jupyter
pytest-cov
pandas
nbval
click
flask
requests
google-cloud-automl
google.cloud